ArchiveBuilder duplicate bug fixed
